+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_createunsecuredattestationtoken.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "9606d67ef2b940110bf088df7d64a58b";
+module.exports.hash = "8596ab4bf4b18fe6bf89c054d04323bc";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}