+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_createunsecuredemptyattestationtoken.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "cbe7176b4c8f0139f445547457173e75";
+module.exports.hash = "1252603f391541dc3e0c8ded369ed525";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}