+++ sdk/attestation/attestation/recordings/node/policygetsettests_/recording_set_policy_sgx__aad_secured.js
@@ -1,0 +1,5 @@
+let nock = require('nock');
+
+module.exports.hash = "99700e41f93237d6d3dd8273c42d8f5f";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}